examples/Python3.12/matplotlib-imageio-example/requirements.txt
contourpy==1.3.3+ppc64le1
cycler==0.12.1
fonttools==4.61.1
ImageIO==2.37.0+ppc64le1
kiwisolver==1.4.4+ppc64le1
matplotlib==3.10.7+ppc64le1
numpy==2.3.4+ppc64le1
packaging==26.0
pillow==12.1.1+ppc64le1
pyparsing==3.2.5+ppc64le1
python-dateutil==2.9.0.post0
setuptools==68.2.2
six==1.17.0
